Fix for msvc
1. msvc requires declare function out of caller function, if
   decalare callee inside caller, callee's namespace is wrong
2. msvc ZLA(zero len array) fields restrictions:
   * disallow fields after ZLA fields
   * disallow deriving from base class which has ZLA field
   * disallow class with ZLA be another class member
3. msvc does not support gcc extension: `res = a ?: b`
4. msvc fiber_aio_read is not implemented, disable it(in MultiGet)
5. msvc set::iterator is not trivially destructable(see MsvcHeap)
6. msvc static_assert(sizeof(A) == xxx) false fail?
7. use NativeOfBigEndian64 instead of __bswap64...
8. use std::_Unsigned128 and relevant changes
9. fix CMakeLists.txt for generate msvc projects
   * disable some msvc warnings
   * generate git-version-*.cc from git-version-each-sideplugin.cc.in
   * many other changes

+++ util/core_local.h
@@ -55,7 +55,7 @@ CoreLocalArray<T>::CoreLocalArray() {
     ++size_shift_;
   }
   size_mask_ = uint16_t((1 << size_shift_) - 1);
-  num_cpus_ = num_cpus;
+  num_cpus_ = uint16_t(num_cpus);
   data_.reset(new T[static_cast<size_t>(1) << size_shift_]);
 }
